I should use this commit function more

+++ analyze_vpp.py
@@ -15,6 +15,7 @@ import shutil
 import scapy.all
 
 INVALID_SPIN_COLOR = "#ff69b450"
+PLOTS_DONE_STRING = "plots_done_1"
 
 ###
 ### Helper functions
@@ -27,7 +28,8 @@ def save_figure(figure, filename):
 	pickle.dump(figure, open("{}.fig.pickle".format(filename), 'wb'))
 	print("Done")
 
-def forward_cursor_to_time(cursor, time, time_series):
+def forward_cursor_t
+impoo_time(cursor, time, time_series):
 	if time_series[0] >= time:
 		return -1
 
@@ -84,13 +86,27 @@ if not os.path.exists("vpp_done"):
 	print("Not analyzed by VPP yet. Goodbye.")
 	sys.exit(1)
 
+if os.path.exists(PLOTS_DONE_STRING):
+	print("Post analysys already done. Goodbye.")
+	sys.exit(1)
+
 ## clean up data from previous runs
 PLOT_DIR = "plots/"
 if os.path.exists(PLOT_DIR):
 	shutil.rmtree(PLOT_DIR)
 os.mkdir(PLOT_DIR)
 
-analyzer_names = ["basic", "pn", "pn_valid", "two_bit", "stat_heur", "rel_heur"]
+## read randID
+
+randID = None
+with open('randID') as randID_file:
+	randID = randID_file.read().strip()
+
+## get directory name
+dir_name = os.path.basename(os.getcwd())
+
+analyzer_names = ["basic", "pn", "pn_valid", "two_bit", "stat_heur", "rel_heur", "handshake"]
+plotable_analyzers = ["basic", "pn", "pn_valid", "two_bit", "stat_heur", "rel_heur"]
 num_of_analyzers = len(analyzer_names)
 
 vpp_data = list()
@@ -108,8 +124,8 @@ with open("switch-2_vpp.csv", newline='') as csvfile:
 
 		vpp_entry = collections.defaultdict(lambda: None)
 		vpp_entry["time"] = float(row["time"]) - base_time
-		vpp_entry["pn"] = int(row["pn"])
-		vpp_entry["host"] = row["host"]
+		vpp_entry["packet_number"] = int(row["pn"])
+		vpp_entry["host"] = row["host"].strip()
 
 		for analyzer in analyzer_names:
 			if row[analyzer + "_new"] == '1':
@@ -121,7 +137,7 @@ with open("switch-2_vpp.csv", newline='') as csvfile:
 ### figure out what the zero epoch time is
 ###
 
-pcap_packets = scapy.all.rdpcap("switch-2_tcpdump.pcap", count=1)
+pcap_packets = scapy.all.rdpcap("switch-2_tcpdump.pcap", count=10)
 zero_epoch = pcap_packets[0].time
 
 print("zero_epoch: {}".format(zero_epoch))
@@ -191,13 +207,26 @@ with open("client-0_ping_stdout.txt") as ping_log:
 			ping_times.append(time)
 
 ###
-### Make plot of all analyzers
+### Find handshake RTT
 ###
-to_plot = analyzer_names
+
+handshake_rtt = None
+for sample in vpp_data:
+	if sample['handshake']:
+		handshake_rtt = sample['handshake']
+		break
+
+with open('plots/handshake_rtt', 'w') as handshake_rtt_file:
+	handshake_rtt_file.write(str(handshake_rtt))
+
+###
+### Make plot of all analyzers together
+###
+to_plot = plotable_analyzers
 f, axarr = plt.subplots(len(to_plot))
 f.set_size_inches(20, 13)
 
-axarr[0].set_title("Comparison of spin analyzers".format())
+axarr[0].set_title("Comparison of spin analyzers [{}]".format(dir_name))
 
 min_x_val = math.inf
 max_x_val = -math.inf
@@ -237,6 +266,40 @@ save_figure(plt.gcf(), PLOT_DIR + "/all_analyzers")
 print([min_x_val, max_x_val])
 
 ###
+### Make plots of all analyzers seperatly
+###
+
+to_plot = plotable_analyzers
+for i in range(len(to_plot)):
+	f, axes = plt.subplots(1)
+
+	analyzer_name = analyzer_names[i]
+
+	axes.set_title("{} [{}]".format(analyzer_name, dir_name))
+
+	y_values_analyzer = [ x[analyzer_name] for x in vpp_data ]
+	x_values_analyzer = [ x["time"] for x in vpp_data ]
+	rejected_x_values = [ x["time"] for x in vpp_data if x[analyzer_name] == None ]
+	rejected_y_values = [ -5 for x in range(len(rejected_x_values)) ]
+
+	client_line = axes.plot(client_times, client_rtts,
+			label="client_estimate", linewidth = .5)
+	server_line = axes.plot(server_times, server_rtts,
+			label="server_estimate", linewidth = .5)
+	ping_line = axes.plot(ping_times, ping_rtts,
+			label="ping", linewidth = .5)
+
+	analyzer_line = axes.plot(x_values_analyzer, y_values_analyzer, label=analyzer_name, linewidth = .5)
+	rejected_marks = axes.plot(rejected_x_values, rejected_y_values, 'rx')
+
+	axes.set_ylim([-10, 120])
+	axes.grid()
+	axes.legend(loc = 2)
+	f.text(0.1, 0.9, "Handshake_rtt: {}".format(handshake_rtt))
+
+	save_figure(plt.gcf(), PLOT_DIR + "/analyzer-" + analyzer_name)
+
+###
 ### Interpollate the client estimates to the sample points of the analyzer
 ###
 
@@ -248,6 +311,8 @@ plt.figure()
 
 plt.plot(client_interpol_times, client_interpol_rtts, 'x', label="interpollated")
 plt.plot(client_times, client_rtts, '*', label="original")
+plt.grid()
+plt.title(dir_name)
 
 
 save_figure(plt.gcf(), PLOT_DIR + "/interpollated_client")
@@ -272,6 +337,8 @@ plt.plot([ x["client"] for x in vpp_data ], [ x["pn_valid"] for x in vpp_data ],
 			'.', markersize=1)
 plt.xlabel("client RTT estimates [ms]")
 plt.ylabel("pn_valid RTT estimates [ms]")
+plt.title(dir_name)
+plt.grid()
 save_figure(plt.gcf(), PLOT_DIR + "/correlation_scatter")
 
 ###
@@ -293,9 +360,9 @@ f.set_size_inches(10, 7)
 x_limits = (-40, 40)
 marker_distance = 1
 
-to_plot = analyzer_names
-markers = ("o", "v", 's', 'p', 'h', '8')
-colors = ('b', 'g', 'r', 'c', 'm', 'y')
+to_plot = plotable_analyzers
+markers = ("o", "v", 's', 'p', 'h', '8', "x")
+colors = ('b', 'g', 'r', 'c', 'm', 'y', "x")
 
 # first plot the lines
 
@@ -320,15 +387,19 @@ for markersonly in range(1): #2):
 				color = colors[i],
 				linestyle = linestyle,
 				linewidth = 1,
-				#marker = markers[i],
+				marker = markers[i],
 				#markevery = (float(i), float(marker_distance * len(to_plot))),
-				#markevery = (0.01*i, 0.1),
-				#markeredgewidth = 1,
-				#markersize = 5,
-				#markeredgecolor = colors[i],
-				#markerfacecolor = (0,0,0,0),
+				markevery = (0.01*i, 0.1),
+				markeredgewidth = 1,
+				markersize = 5,
+				markeredgecolor = colors[i],
+				markerfacecolor = (0,0,0,0),
 				label = label)
 
 plt.xlim(x_limits)
 plt.legend()
+plt.title(dir_name)
+plt.grid()
 save_figure(plt.gcf(), PLOT_DIR + "/ECDF")
+
+open(PLOTS_DONE_STRING, 'w').close()
